API update definitivo - banco-model: necesita informacion para hacer datos persistentes - banco-ejb: necesita informacion para comunicarse con banco-web y banco-model

+++ banco/banco-ejb/src/main/java/es/uc3m/tiw/ejb/LogicaPedidos.java
@@ -6,6 +6,6 @@ public interface LogicaPedidos {
 	
 	String validarPedido(Pedido pedido);
 	
-	Integer conciliarPedido(String cod_pedido);
+	Double conciliarPedido(String cod_pedido);
 
 }
